--- src/app/whenToMeet/page.module.css
@@ -1,13 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100vh;
-  /* justify-content: center;
-  margin-top: 7%;
-  margin-bottom: 7%; */
-}
-
-.desc {
-  margin: 2rem 0;
-}
